- fixx issues like timestamp sorting, now based correctly on time since epoch with an offset to sorrectly sort new likes but keep old ones in their previous order, also switches to using f64 for all timestamps to allow enough precision

+++ app/code/main.h
@@ -307,7 +307,7 @@ struct ReleasesView
 struct ChartItem
 {
     std::string index;
-    u32         pos;
+    f64         pos;
 };
 
 struct Store
